requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
openai==1.3.7
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv @ git+https://github.com/theskumar/python-dotenv@3ffcef60d10813b72ecf85d5941d51b0207cd40e
pytz==2023.3.post1
pyzmq==25.1.2
reportlab==4.0.7
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1